requirements.txt
azure-functions
pandas
gremlinpython
SPARQLWrapper
msrest
azure-eventgrid